+++ RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 This document contains high-level release notes. More details can be found in [milestones](https://github.com/searchisko/searchisko/issues/milestones) on GitHub.
 
+## 2.0.1-SNAPSHOT
+
+- TBD
+
 ## 2.0.0
 
 Breaking: